promote all handle type to 64bit unsigned

+++ cbits/proAudio.cpp
@@ -146,14 +146,14 @@ DeviceAudio::DeviceAudio() : m_freqOut(0), m_volL(1.0f), m_volR(1.0f) {
     loaderRegister(loadOgg,"ogg");
 }
 
-unsigned int DeviceAudio::sampleFromFile(const std::string & filename, float volume) {
+uint64_t DeviceAudio::sampleFromFile(const std::string & filename, float volume) {
     if(filename.rfind('.')>filename.size()) return 0;
     string suffix=toLower(filename.substr(filename.rfind('.')+1));
     map<string, AudioSample * (*)(const string &)>::iterator it = mm_loader.find(suffix);
     if(it==mm_loader.end()) return 0;
     AudioSample* pSample = (*(it->second))(filename);
     if(!pSample) return 0;
-    unsigned int ret = sampleFromMemory(*pSample, volume);
+    uint64_t ret = sampleFromMemory(*pSample, volume);
     delete pSample;
     return ret;
 }
